cleanup, renames, dynamic message subject
