+++ app/email_templates/confirmation.js
